fix: 빌드에러 해결 및 "찜한 모임"페이지 파일 이동 - 헤더 컴포넌트 바뀐 코드로 변경 해서 빌드오류 해결 - "찜한 모임"페이지 liked 폴더에서 savedGathering폴더로 이동
